feat(gatsby-core-utils): Add file download functions
