+++ packages/gatsby-core-utils/package.json
@@ -31,6 +31,7 @@
   "dependencies": {
     "ci-info": "2.0.0",
     "configstore": "^5.0.1",
+    "file-type": "^16.2.0",
     "fs-extra": "^8.1.0",
     "node-object-hash": "^2.0.0",
     "proper-lockfile": "^4.1.1",
