+++ packages/gatsby-core-utils/src/index.ts
@@ -7,5 +7,6 @@ export { createRequireFromPath } from "./create-require-from-path"
 export { getConfigStore } from "./get-config-store"
 export { getGatsbyVersion } from "./get-gatsby-version"
 export { getTermProgram } from "./get-term-program"
+export { fetchRemoteFile } from "./fetch-remote-file"
 export * from "./service-lock"
 export * from "./site-metadata"
