+++ packages/gatsby-source-filesystem/src/utils.js
@@ -1,6 +1,5 @@
 const path = require(`path`)
 const Url = require(`url`)
-const ProgressBar = require(`progress`)
 
 /**
  * getParsedPath
